Rename module that better shows the intent
