+++ homeassistant/components/ping/helpers.py
